# Remove type-where in paginated queries